Integration of DMC, SSP Grid generation is now functional for all models

+++ src/DDRep_cpp.h
@@ -7,7 +7,7 @@
 
 struct RAW_format
 {
-  long time;
+  double time;
   int  resp;
   std::string  cond;
   bool operator<(const RAW_format& rhs) const
